Close #11; restructure IO handling and data stored in shapes

+++ src/net/premiumrich/shapes/EllipseShape.java
@@ -4,7 +4,7 @@ import java.awt.geom.Ellipse2D;
 
 public class EllipseShape extends MapShape {
 
-	public EllipseShape(double x, double y, int width, int height) {
+	public EllipseShape(int x, int y, int width, int height) {
 		super(new Ellipse2D.Double(x, y, width, height), x, y);
 	}
